+++ mythtv/programs/mythtranscode/external/replex/element.cpp
@@ -211,7 +211,7 @@ int get_video_info(ringbuffer *rbuf, sequence_t *s, int off, int le)
 	s->h_size	= ((headr[1] &0xF0) >> 4) | (headr[0] << 4);
 	s->v_size	= ((headr[1] &0x0F) << 8) | (headr[2]);
     
-        int sw = (int)((headr[3]&0xF0) >> 4) ;
+        int sw = ((headr[3]&0xF0) >> 4) ;
 
 	if (DEBUG){
 		switch( sw ){
